Added Intro Animation Sequence
Video animation sequence has been added. Scene name is AnimationSetup. Asset name is Al_FrightNight_AnimationSequence_FinalColor.MP4 and has been added in the scene. I think some coding is necessary for timing when the clip plays, but all I had to do to get it to play was literally just to import the video into the scene and press the play button.