Intro Animation Added
Added Intro Animation of the child hiding in the clothes as well as a barebones script intended to trigger the cutscene when the player reaches a certain area. Again, please edit as you guys see fit, I'm not very experienced at programming in Unity, so there are definitely errors in the code.